Introducing the Contributors Bounty Board!
In this update, I have added "contribute.md" to the docs directory of this repository.  The document highlights how contributions to the project can be done, and also introduces the Contributors Bounty Board!

The Contributors Bounty Board lists certain features/things that need to be done/fixed for VisorWare to continue development.

You can find the board in the Projects Tab of the GitHub Repository. Please read through docs/contribute.md for more info!